Класс Node определён в классе InMemoryHistoryManager
