+++ src/manager/InMemoryHistoryManager.java
@@ -1,6 +1,5 @@
 package manager;
 
-import task.Node;
 import task.Task;
 
 import java.util.List;
@@ -75,5 +74,41 @@ public class InMemoryHistoryManager implements HistoryManager {
             arrTask.add(node.getData());
         }
         return arrTask;
+    }
+
+    public static class Node {
+        private Task data;
+        private Node next;
+        private Node prev;
+
+        public Node(Node prev, Task data, Node next) {
+            this.data = data;
+            this.next = next;
+            this.prev = prev;
+        }
+
+        public Task getData() {
+            return data;
+        }
+
+        public void setData(Task data) {
+            this.data = data;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
     }
 }
